Fix a missing import in RN debugger support.
