+++ lib/browser/index.js
@@ -31,6 +31,7 @@ import Credentials, { createCredentials } from './credentials';
 import * as rpc from './rpc';
 import * as util from './util';
 import { static as staticUserMethods } from '../user';
+import { createSession } './session';
 
 const {debugHosts, debugPort} = NativeModules.Realm;
 
@@ -210,9 +211,9 @@ Object.defineProperties(Realm, {
 
 for (let i = 0, len = debugHosts.length; i < len; i++) {
     try {
-        let refreshAccessTokenCallback = staticUserMethods._refreshAccessToken.bind(User);
+        let refreshAccessTokenCallback = staticUserMethods && staticUserMethods._refreshAccessToken ? staticUserMethods._refreshAccessToken.bind(User) : () => {};
         // The session ID refers to the Realm constructor object in the RPC server.
-        Realm[keys.id] = rpc.createSession(refreshAccessTokenCallback, debugHosts[i] + ':' + debugPort);
+        Realm[keys.id] = createSession(refreshAccessTokenCallback, debugHosts[i] + ':' + debugPort);
         break;
     } catch (e) {
         // Only throw exception after all hosts have been tried.
