+++ lib/browser/rpc.js
@@ -21,6 +21,7 @@
 import * as base64 from './base64';
 import * as util from './util';
 import { keys, objectTypes } from './constants';
+import { invalidateCache } from './cache';
 
 const { id: idKey, realm: _realmKey } = keys;
 let registeredCallbacks = [];
@@ -61,7 +62,7 @@ function beforeNotify(realm) {
 
     // Clear the cache for this Realm, and reenable caching if it was disabled
     // by a write transaction.
-    util.invalidateCache(realm[keys.realm]);
+    invalidateCache(realm[keys.realm]);
 }
 
 export function createRealm(args) {
