+++ lib/browser/session.js
@@ -53,10 +53,12 @@ export function createSession(realmId, info) {
     // the realm id to be present on any object that is used over rpc
     sessionProxy[keys.realm] = "(Session object)";
 
-    sessionProxy[keys.id] = info.id;
-    sessionProxy[keys.type] = objectTypes.SESSION;
-    sessionProxy.user = deserialize(realmId, info.data.user);
-    sessionProxy.config = deserialize(realmId, info.data.config)
+    if (info && info.data && info.data.user) {
+        sessionProxy[keys.id] = info.id;
+        sessionProxy[keys.type] = objectTypes.SESSION;
+        sessionProxy.user = deserialize(realmId, info.data.user);
+        sessionProxy.config = deserialize(realmId, info.data.config);
+    }
 
     return sessionProxy;
 }
