+++ lib/browser/util.js
@@ -20,6 +20,7 @@
 
 import { keys } from './constants';
 import * as rpc from './rpc';
+import {invalidateCache, getRealmCache} from './cache';
 
 export function createMethods(prototype, type, methodNames, mutating) {
     let props = {};
@@ -57,25 +58,6 @@ export function createMethod(type, name, mutating) {
             }
         }
     };
-}
-
-let propertyCache = {};
-
-export function invalidateCache(realmId) {
-    if (realmId) {
-        propertyCache[realmId] = {};
-    }
-    else {
-        propertyCache = {};
-    }
-}
-
-function getRealmCache(realmId) {
-    let realmCache = propertyCache[realmId];
-    if (!realmCache) {
-        realmCache = propertyCache[realmId] = {};
-    }
-    return realmCache;
 }
 
 export function cacheObject(realmId, id, value) {
